click-me: introduce loader for training set
