+++ harmonization/__init__.py
@@ -11,3 +11,4 @@ __version__ = '0.0.4'
 from . import evaluation
 from . import training
 from . import models
+from . import common
