+++ harmonization/evaluation/__init__.py
@@ -2,5 +2,5 @@
 Module related to the Human Alignment benchmark
 """
 
-from .click_me import load_clickme_val, evaluate_clickme
+from .clickme import load_clickme_val, evaluate_clickme
 from .metrics import dice, spearman_correlation, intersection_over_union
